I removed the Order::$url property as it seems to no longer be needed.
Then I renamed the $urlHelper property to $url (it was only named that
originally due to the naming conflict, which is now gone).

ALSO, it turns out that knockout.js attaches an event to links on the
orders page, such that when you left-click on a link it takes you to the
(Magento) order page, regarless of the href. I had foolishly tested this
originally by examining the link url in the status bar and right-clicking
the link to open the iframe in a new tab (which works fine).

Yes, that's right, when you left-click the link, Magento takes you to a
COMPLETELY DIFFERENT URL than when you right-click that same link.